# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Simple Light
  description: "Simple application (OTP) for API course at University of Szeged"
  contact: { 
    "name": "Berki Tamás",
    "email": "Berki.Tamas.Jozsef@stud.u-szeged.hu",
    "url": "http://btj.hu"
  }
  license: {
    "name": "MIT",
    "url": "https://github.com/berkitamas/simplelight/blob/master/LICENSE"
  }
host: localhost:10010
basePath: /
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
...
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: header
    name: X-Session-ID
security:
  - ApiKeyAuth: []
paths:
...
  /users/{userId}/card:
    x-swagger-router-controller: cards
    parameters:
      - name: userId
        in: path
        required: true
        description: ID of the user
        type: string
    get:
      summary: Retrieve all card of user
      description: Returns all card of a user
      operationId: retrieveAllCard
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/Card"
        400:
          description: Request validation failed
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: User is not logged in
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: User with given ID not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      summary: Add card to user
      description: Adds a card to the user account
      operationId: addCard
      parameters:
        - name: card
          in: body
          description: Details of the card
          schema:
            $ref: "#/definitions/Card"
      responses:
        201:
          description: Successfully created
          schema:
            $ref: "#/definitions/Card"
        400:
          description: Request validation failed
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: User is not logged in
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: User with given ID not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/{userId}/cards/{cardId}:
    x-swagger-router-controller: cards
    parameters:
      - name: userId
        in: path
        required: true
        description: ID of the user
        type: string
      - name: cardId
        in: path
        required: true
        description: ID of the card
        type: string
    get:
      tags:
      - card
      - user
      summary: Retrieve card
      description: Returns the card of the given id
      operationId: retrieveCard
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/Card"
        400:
          description: Request validation failed
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: User is not logged in
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: User or card with given ID not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      tags:
        - card
        - user
      summary: Update card details
      description: Updates the card of the given id
      operationId: updateCard
      parameters:
        - name: card
          in: body
          required: true
          description: Details of the new card
          schema:
            $ref: "#/definitions/Card"
      responses:
        200:
          description: Successfully changed
          schema:
            $ref: "#/definitions/Card"
        400:
          description: Request validation failed
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: User is not logged in
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: User or card with given ID not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: deleteCard
      tags:
        - card
      summary: Delete card
      description: Deletes the card of the given id
      responses:
        204:
          description: Successfully deleted
        400:
          description: Request validation failed
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: User is not logged in
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: User or card with given ID not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
          
  # Services #
  /users/{userId}/motorwayTickets:
    x-swagger-router-controller: motorways
    parameters:
      - name: userId
        in: path
        required: true
        description: ID of the user
        type: string
    get:
      tags:
        - motorway
        - user
      summary: Retrieve all motorway ticket
      description: Return all motorway tickest (past one also) of the given user
      operationId: retrieveAllMotorwayTicket
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/MotorwayTicket"
        400:
          description: Request validation failed
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: User is not logged in
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: User with given ID not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      tags:
        - motorway
        - user
      summary: Add motorway ticket
      description: Adds a new motorway ticket for the given user
      operationId: addMotorwayTicket
      parameters:
        - name: ticket
          in: body
          required: true
          description: Details of the ticket
          schema:
            $ref: "#/definitions/MotorwayTicket"
      responses:
        201:
          description: Successfully created
        400:
          description: Request validation failed
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: User is not logged in
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: User with given ID not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/{userId}/motorwayTickets/{ticketId}:
    x-swagger-router-controller: motorways
    parameters:
      - name: ticketId
        in: path
        required: true
        description: ID of the ticket
        type: string
      - name: userId
        in: path
        required: true
        description: ID of the user
        type: string
    get:
      tags:
        - motorway
        - user
      summary: Retrieve a motorway ticket
      description: Retrieves tha motorway ticket of the user with the given ids
      operationId: retrieveMotorwayTicket
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/MotorwayTicket"
        400:
          description: Request validation failed
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: User is not logged in
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: User with given ID not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /users/{userId}/cinemaTickets:
    x-swagger-router-controller: cinemas
    parameters:
      - name: userId
        in: path
        required: true
        description: ID of the user
        type: string
    get:
      tags:
        - cinema
      summary: Retrieve all cinema tickets
      description: Returns all cinema tickets of the given user
      operationId: retrieveAllCinemaTicket
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/CinemaTicket"
        400:
          description: Request validation failed
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: User is not logged in
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: User with given ID not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      tags:
        - cinema
      summary: Add cinema ticket
      description: Creates a cinema ticket for the user
      operationId: addCinemaTicket
      parameters:
        - name: ticket
          in: body
          required: true
          description: Details of the cinema ticket
          schema:
            $ref: "#/definitions/CinemaTicket"
      responses:
        200:
          description: OK
        400:
          description: Request validation failed
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: User is not logged in
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: User with given ID not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/{userId}/cinemaTickets/{ticketId}:
    x-swagger-router-controller: cinemas
    parameters:
      - name: ticketId
        in: path
        required: true
        description: ID of the ticket
        type: string
      - name: userId
        in: path
        required: true
        description: ID of the user
        type: string
    get:
      tags:
        - cinema
      summary: Retrieve cinema ticket
      description: Returns the cinema ticket from the given user and ticket ID
      operationId: retrieveCinemaTicket
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/CinemaTicket"
        400:
          description: Request validation failed
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: User is not logged in
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: User or ticket with given ID not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - cinema
      summary: Delete cinema ticket
      description: Deletes cinema ticket of the given user
      operationId: deleteCinemaTicket
      responses:
        204:
          description: Successfully deleted
        400:
          description: Request validation failed
          schema:
            $ref: "#/definitions/ErrorResponse"
        403:
          description: User is not logged in
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: User or ticket with given ID not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /cinemas:
    x-swagger-router-controller: cinemas
    get:
      tags:
        - cinema
      security: []
      summary: Retrieve all cinema
      description: Returns all of the cinemas
      operationId: retrieveAllCinema
      parameters:
        - name: city
          in: query
          required: false
          type: string
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/CinemaSummary"
  /cinemas/{cinemaId}:
    x-swagger-router-controller: cinemas
    parameters:
      - name: cinemaId
        in: path
        required: true
        description: ID of the cinema
        type: string
    get:
      tags:
        - cinema
      security: []
      summary: Retrieve all show in cinema
      description: Returns all of the shows in a given cinema
      operationId: retrieveAllShowInCinema
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/Cinema"
        404:
          description: Cinema not found
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  Card:
    required:
      - cardholderName
      - cardNumber
      - expirationMonth
      - expirationYear
      - cvv
    properties:
      _id:
        type: string
      userId:
        type: string
      bank:
        type: string
      cardholderName:
        type: string
      cardNumber:
        type: string
      expirationMonth:
        type: string
      expirationYear:
        type: string
      cvv:
        type: string
  MotorwayTicket:
    required:
      - cardId
      - licensePlate
      - priceCategory
    properties:
      _id:
        type: string
      userId:
        type: string
      cardId:
        type: string
      licensePlate:
        type: string
      priceCategory:
        type: string
        enum: [Weekly, Monthly, Yearly]
      purchaseDate:
        type: string
        format: date-time
...
  CinemaTicket:
    required:
      - cardId
      - showId
      - row
      - seat
    properties:
      _id:
        type: string
      userId:
        type: string
      cardId:
        type: string
      showId:
        type: string
      row:
        type: number
      seat:
        type: number
      purchaseDate:
        type: string
        format: date-time
...
  CinemaSummary:
    properties:
      _id:
        type: string
      cinemaName:
        type: string
      zip:
        type: string
      city: 
        type: string
      street:
        type: string
      address:
        type: string
  Cinema:
    properties:
      summary:
        $ref: "#/definitions/CinemaSummary"
      shows:
        type: array
        items:
          $ref: "#/definitions/CinemaShow"
  Movie:
    properties:
      name:
        type: string
      summary:
        type: string
      coverUrl:
        type: string
      imdbLink:
        type: string
  CinemaShow:
    properties:
      _id:
        type: string
      startTime:
        type: string
        format: date-time
      movie:
        $ref: "#/definitions/Movie"
      price:
        type: string
      hall:
        type: string
      maxRows:
        type: number
      maxSeatsPerRow:
        type: number
      freePositions:
        type: array
        items:
          $ref: "#/definitions/CinemaShowSeat"
  CinemaShowSeat:
    properties:
      row:
        type: number
      seat:
        type: number
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string